Refactor: consolidate common code into programmatic_usage_common.ts
- Rename programmatic_usage_es_filter.ts to programmatic_usage_common.ts
- Move USAGE_ORIGINS_CLASSIFICATION to common (remove duplicate)
- Move USER_USAGE_ORIGINS to common
- Move MARKUP_MULTIPLIER to common (shared by daily_cap and key_cap)
- Move getSecondsUntilMidnightUTC to common (shared by daily_cap and key_cap)
- Move UsageAggregations type to common
- Update all imports to use common directly (no re-exports)
